+++ tomcat/src/main/java/org/apache/catalina/controller/AbstractController.java
@@ -1,5 +1,6 @@
-package org.apache.coyote.controller;
+package org.apache.catalina.controller;
 
+import org.apache.catalina.Controller;
 import org.apache.coyote.http11.message.HttpRequest;
 import org.apache.coyote.http11.message.HttpResponse;
